Make trainer.py train by overselection, take 1
In theory, modify trainer.py to train via overselection rather than just dropping the bottom half of the trees. In practice, this may well not actually work correctly, and I have a bunch of homework to do instead of testing it. Decent odds something's shallowly wrong here, and it's entirely possible something's deeply wrong. We'll see.