Added in the old training method
Old training method now available under old_train. Works as in the master branch, only the ratio of selected bots has an input that can be changed. Doesn't actually work if changed, though, because that would need us to reuse the selected bots and there isn't code for that right now. Contains commented out code that might solve that problem, but it's untested.